+++ school/src/app/config.js
@@ -12,7 +12,7 @@ module.exports = {
   MYSQL_PREFIX,
   MYSQL_INSTALL,
   CONNECTIONLIMIT,
-  DEFAULT_ADMIN_USERNAME,
-  DEFAULT_ADMIN_ACCOUNT,
-  DEFAULT_ADMIN_PASSWORD
+  // DEFAULT_ADMIN_USERNAME,
+  // DEFAULT_ADMIN_ACCOUNT,
+  // DEFAULT_ADMIN_PASSWORD
 } = process.env;
