+++ school/src/main.js
@@ -1,4 +1,3 @@
-/* 版权所有 ©2022-2025 QQ：1440272370 该管理系统遵守MIT开源协议 */
 const app = require('./app');
 const { APP_PORT } = require('./app/config');
 require('./app/database');
